requirements.txt
torch==2.3.1
numpy
pandas
datasets
evaluate
transformers
sentencepiece
faiss-cpu
sqlite-utils
pyyaml
datasketch
tiktoken
tqdm
pytest
rich
tabulate
